requirements.txt
tqdm
numpy
cmake
boost
pillow
matplotlib
scikit-image
torch
torchvision
opencv-python